Uses class ConfigForm in ActivityPanel
