+++ frog/imports/ui/GraphEditor/SidePanel/FormUtils/addSocialSchema.js
@@ -1,4 +1,5 @@
 // @flow
+
 import { cloneDeep, set, get, merge } from 'lodash';
 import traverse from 'traverse';
