+++ frog/imports/ui/GraphEditor/SidePanel/FormUtils/index.js
@@ -3,3 +3,4 @@
 export { default as SelectActivityWidget } from './SelectActivity';
 export { default as SelectAttributeWidget } from './SelectAttribute';
 export { default as SelectFormWidget } from './SelectForm';
+export { default as addSocialFormSchema } from './addSocialSchema';
